[FIX] changed structure
