+++ src/apis/category/CategoryDto.ts
@@ -14,3 +14,12 @@ export interface UpdateCategoryResponse {
   index: number
   name: string
 }
+
+export interface CreateCategoryRequest {
+  masterCategoryId: number | null
+  categoryName: string
+}
+
+export interface UpdateCategoryRequest {
+  name: string
+}
