--- src/apis/category/dto/CategoryRequest.ts
@@ -1,8 +1,0 @@
-export interface CreateCategoryRequest {
-  masterCategoryId: number | null
-  categoryName: string
-}
-
-export interface UpdateCategoryRequest {
-  name: string
-}
